Silence nbtree.h cpluspluscheck warning.
Add a cast to size_t to silence "comparison between signed and unsigned
integer expressions" cpluspluscheck warning.

Reported-By: Tom Lane
Discussion: https://postgr.es/m/7971.1583171266@sss.pgh.pa.us

+++ src/include/access/nbtree.h
@@ -381,7 +381,7 @@ static inline void
 BTreeTupleSetPosting(IndexTuple itup, int nhtids, int postingoffset)
 {
 	Assert(nhtids > 1 && (nhtids & BT_OFFSET_MASK) == nhtids);
-	Assert(postingoffset == MAXALIGN(postingoffset));
+	Assert((size_t) postingoffset == MAXALIGN(postingoffset));
 	Assert(postingoffset < INDEX_SIZE_MASK);
 
 	itup->t_info |= INDEX_ALT_TID_MASK;
